Add `Software` attribute to relay messages
